Added isAlive bool to EntityAI.cs Added cleanupOnDeath bool to EntityAI.cs
Many of null reference exceptions the AI was throwing was from the AI still trying to do stuff after it was dead. All Automated processes now operate under the condition that it is alive.

On death, the AI script members are properly cleaned up, destroyed, or deactivated exactly once.

Known Bugs:
test_Dwarf.prefab and test_Orgling.prefab were causing on multiple triggers. I added an extra condition to not update the list if the character is already in the list.

Moving Forward
Clean up what I have, because everything starting to look crazy.